✨ ♻️  BSTNode Fixes, KeySet Method, Main code, Refactoring
Added and tested key set method that relies on ordered list of keys. Refactored some of the code in treemap so it is more efficient, such as using values instead of references. Added main to test out with car values.  For BSTNode fixed code based on email. BinaryTree map was missing post order, preorder and in order code so I added it.