✨ Weapon class, CSV read method, simple main implementation
Created a weapon class which holds a struct for each weapon from the csv file. Added a method that makes it able to read from the csv file, using stringstream to ignore the commas. I then implemented a simple main that prints out each item to check if it works.